feature #10: 각 모듈 생성 및 연결
