+++ Projects/Domain/Sources/Example.swift
@@ -1,0 +1,2 @@
+import Foundation
+
