+++ Projects/Keyme/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
