+++ Projects/Keyme/Tests/Sources/ExampleTests.swift
@@ -1,0 +1,8 @@
+import Foundation
+import XCTest
+
+final class ExampleTests: XCTestCase {
+    func test_twoPlusTwo_isFour() {
+        XCTAssertEqual(2+2, 4)
+    }
+}
