--- Targets/Keyme/Sources/KeymeApp.swift
@@ -1,53 +1,0 @@
-import KeymeUI
-import SwiftUI
-import UserNotifications
-
-import FirebaseCore
-import FirebaseMessaging
-
-@main
-struct KeymeApp: App {
-    @UIApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
-
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-    }
-}
-
-class AppDelegate: UIResponder, UIApplicationDelegate {
-    func application(
-        _ application: UIApplication,
-        didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
-    ) -> Bool {
-        FirebaseApp.configure()
-        Messaging.messaging().delegate = self
-
-        UNUserNotificationCenter.current().requestAuthorization(options: [.alert, .badge, .sound]) { granted, _ in
-            guard granted else { return }
-            
-            // 푸시토큰 애플 서버에 등록하기
-            self.getNotificationSettings()
-        }
-        return true
-    }
-}
-
-extension AppDelegate: MessagingDelegate {
-    func messaging(_ messaging: Messaging, didReceiveRegistrationToken fcmToken: String?) {
-        // Do something
-        print(messaging, fcmToken as Any)
-    }
-}
-
-extension AppDelegate {
-    private func getNotificationSettings() {
-        UNUserNotificationCenter.current().getNotificationSettings { settings in
-            guard settings.authorizationStatus == .authorized else { return }
-            DispatchQueue.main.async {
-                UIApplication.shared.registerForRemoteNotifications()
-            }
-        }
-    }
-}
