--- Targets/Keyme/Tests/KeymeTests.swift
@@ -1,8 +1,0 @@
-import Foundation
-import XCTest
-
-final class KeymeTests: XCTestCase {
-    func test_twoPlusTwo_isFour() {
-        XCTAssertEqual(2+2, 4)
-    }
-}
