--- Targets/KeymeKit/Tests/KeymeKitTests.swift
@@ -1,8 +1,0 @@
-import Foundation
-import XCTest
-
-final class KeymeKitTests: XCTestCase {
-    func test_example() {
-        XCTAssertEqual("KeymeKit", "KeymeKit")
-    }
-}
