--- Targets/KeymeUI/Tests/KeymeUITests.swift
@@ -1,8 +1,0 @@
-import Foundation
-import XCTest
-
-final class KeymeUITests: XCTestCase {
-    func test_example() {
-        XCTAssertEqual("KeymeUI", "KeymeUI")
-    }
-}
